First commit of hexagon metasurface
This class defines a moving metasurface composed entirely of hexagon objects. There is no rotation, so each is defined by only three parameters (x,y,r) with r being the radius of the circle in which the hexagons are inscribed. (These are regular hexagons). This draft of the code has a working geometry import but no gradient calculation or plotting.